// types.ts
export type Difficulty = 'Iniciante' | 'Intermediário' | 'Avançado';
export type Category = 'Aquecimento' | 'Pé' | 'Sentado' | 'Inversão' | 'Restaurativa' | 'Core' | 'Finalização' | 'Respiração';
export type Goal = 'Flexibilidade' | 'Força' | 'Relaxamento' | 'Alívio de Dor';
export type Duration = 15 | 30 | 45;
export type Discomfort = 'Lombar' | 'Joelhos' | 'Pescoço/Ombros' | 'Punhos' | 'Nenhum';

export interface User {
  id: string;
  name: string;
  email: string;
  isAdmin?: boolean;
}

export interface Media {
  thumbnailUrl: string;
  videoEmbedUrl: string;
}

export interface Pose {
  id: string;
  sanskritName: string;
  portugueseName: string;
  difficulty: Difficulty;
  category: Category;
  benefits: string[];
  media: Media;
  durationDefault: number; // in seconds
  description: string;
}
// ...
export interface Routine {
  id: string;
  name: string;
  poses: Pose[];
  totalDuration: number;
  createdAt: Date;
}